Properly fix filling of text properties onto following line
If erc-fill or erc-insert-timestamp would flow the line onto a subsequent
line then the text properties of the last character we had inserted would
get inherited across this new line. As such, the underline etc of
erc-button-face would then be applied and we would get lines like:

<alexmurray> yep, it's working now. It definitely wasn't working earlier and all weekend though. [00a]_
___________________________________________________________________________________________________ [13:38]

Instead this is fixed so we see:

<alexmurray> yep, it's working now. It definitely wasn't working earlier and all weekend though. [00a]
                                                                                                    [13:38]

As expected.